+++ ourJourney/src/main/java/pudding/toy/ourJourney/comment/service/CommentService.java
@@ -5,10 +5,10 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import pudding.toy.ourJourney.comment.CommentRepository;
 import pudding.toy.ourJourney.comment.dto.CommentProfileDto;
 import pudding.toy.ourJourney.comment.dto.GetCommentsDto;
 import pudding.toy.ourJourney.comment.entity.Comment;
+import pudding.toy.ourJourney.comment.repository.CommentRepository;
 import pudding.toy.ourJourney.content.entity.Contents;
 import pudding.toy.ourJourney.content.repository.ContentRepository;
 import pudding.toy.ourJourney.profile.entity.Profile;
