+++ ourJourney/src/main/java/pudding/toy/ourJourney/config/DummyDataInitializer.java
@@ -7,8 +7,8 @@ import org.springframework.core.env.Profiles;
 import org.springframework.stereotype.Component;
 import pudding.toy.ourJourney.category.entity.Category;
 import pudding.toy.ourJourney.category.repository.CategoryRepository;
-import pudding.toy.ourJourney.comment.CommentRepository;
 import pudding.toy.ourJourney.comment.entity.Comment;
+import pudding.toy.ourJourney.comment.repository.CommentRepository;
 import pudding.toy.ourJourney.content.entity.Contents;
 import pudding.toy.ourJourney.content.repository.ContentRepository;
 import pudding.toy.ourJourney.profile.entity.Profile;
