+++ ourJourney/src/main/java/pudding/toy/ourJourney/profile/service/ProfileService.java
@@ -10,8 +10,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 import pudding.toy.ourJourney.auth.dto.ProfileAuthRequest;
-import pudding.toy.ourJourney.comment.CommentRepository;
 import pudding.toy.ourJourney.comment.entity.Comment;
+import pudding.toy.ourJourney.comment.repository.CommentRepository;
 import pudding.toy.ourJourney.content.entity.ContentLike;
 import pudding.toy.ourJourney.content.entity.Contents;
 import pudding.toy.ourJourney.content.repository.ContentLikeRepository;
