+++ ourJourney/src/test/java/pudding/toy/ourJourney/service/CommentServiceTest.java
@@ -7,9 +7,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
-import pudding.toy.ourJourney.comment.CommentRepository;
 import pudding.toy.ourJourney.comment.dto.GetCommentsDto;
 import pudding.toy.ourJourney.comment.entity.Comment;
+import pudding.toy.ourJourney.comment.repository.CommentRepository;
 import pudding.toy.ourJourney.comment.service.CommentService;
 import pudding.toy.ourJourney.content.entity.Contents;
 import pudding.toy.ourJourney.content.repository.ContentRepository;
