+++ ourJourney/src/test/java/pudding/toy/ourJourney/service/ContentServiceTest.java
@@ -7,8 +7,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 import pudding.toy.ourJourney.category.entity.Category;
 import pudding.toy.ourJourney.category.repository.CategoryRepository;
-import pudding.toy.ourJourney.comment.CommentRepository;
 import pudding.toy.ourJourney.comment.entity.Comment;
+import pudding.toy.ourJourney.comment.repository.CommentRepository;
 import pudding.toy.ourJourney.content.dto.DetailContentResponse;
 import pudding.toy.ourJourney.content.entity.ContentLike;
 import pudding.toy.ourJourney.content.entity.Contents;
